+++ cirq/devices/noise_model_test.py
@@ -19,6 +19,7 @@ import pytest
 
 import cirq
 from cirq import ops
+from cirq.devices.noise_model import validate_all_measurements
 
 
 def _assert_equivalent_op_tree(x: cirq.OP_TREE, y: cirq.OP_TREE):
@@ -199,3 +200,42 @@ def test_gate_substitution_noise_model():
     rho1 = cirq.final_density_matrix(circuit, noise=noise)
     rho2 = cirq.final_density_matrix(circuit2)
     np.testing.assert_allclose(rho1, rho2)
+
+
+def test_moment_is_measurements():
+    q = cirq.LineQubit.range(2)
+    circ = cirq.Circuit([cirq.X(q[0]), cirq.X(q[1]), cirq.measure(*q, key='z')])
+    assert not validate_all_measurements(circ[0])
+    assert validate_all_measurements(circ[1])
+
+
+def test_moment_is_measurements_mixed1():
+    q = cirq.LineQubit.range(2)
+    circ = cirq.Circuit(
+        [
+            cirq.X(q[0]),
+            cirq.X(q[1]),
+            cirq.measure(q[0], key='z'),
+            cirq.Z(q[1]),
+        ]
+    )
+    assert not validate_all_measurements(circ[0])
+    with pytest.raises(ValueError) as e:
+        validate_all_measurements(circ[1])
+    assert e.match(".*must be homogeneous: all measurements.*")
+
+
+def test_moment_is_measurements_mixed2():
+    q = cirq.LineQubit.range(2)
+    circ = cirq.Circuit(
+        [
+            cirq.X(q[0]),
+            cirq.X(q[1]),
+            cirq.Z(q[0]),
+            cirq.measure(q[1], key='z'),
+        ]
+    )
+    assert not validate_all_measurements(circ[0])
+    with pytest.raises(ValueError) as e:
+        validate_all_measurements(circ[1])
+    assert e.match(".*must be homogeneous: all measurements.*")
